rename AdminVenues and make it base component
