+++ src/components/organisms/AdminDashboard/index.ts
@@ -1,0 +1,1 @@
+export { AdminDashboard } from "./AdminDashboard";
