--- src/components/organisms/AdminVenues/index.ts
@@ -1,1 +1,0 @@
-export { AdminVenues } from "./AdminVenues";
